+++ VRDR.Tests/MortalityData_Should.cs
@@ -71,7 +71,6 @@ namespace VRDR.Tests
             DeathRecord dr = ije1.ToDeathRecord();
             Dictionary<string, string> age = new Dictionary<string, string>();
             age.Add("value", "10");
-            age.Add("unit", "Months");
             age.Add("code", "mo");
             dr.AgeAtDeath = age;
             Assert.Equal("mo", dr.AgeAtDeath["code"]);
@@ -402,6 +401,19 @@ namespace VRDR.Tests
             ije.DMIDDLE = "Dmiddle";
             Assert.Equal("D", ije.MNAME.Trim());
             Assert.Equal("Dmiddle", ije.DMIDDLE.Trim());
+        }
+
+        // If the DSTATE is not set we should get a different message from it being set to an invalid value
+        [Fact]
+        public void DSTATE_Errors()
+        {
+            DeathRecord record = new DeathRecord();
+            record.DeathLocationJurisdiction = ""; // No jurisdiction code
+            ArgumentOutOfRangeException e = Assert.Throws<ArgumentOutOfRangeException>(() => new IJEMortality(record));
+            Assert.Equal("Specified argument was out of the range of valid values. (Parameter 'Found 1 validation errors:\nError: FHIR field DeathLocationJurisdiction is blank, which is invalid for IJE field DSTATE.')", e.Message);
+            record.DeathLocationJurisdiction = "QQ"; // Not a valid jurisdiction code
+            e = Assert.Throws<ArgumentOutOfRangeException>(() => new IJEMortality(record));
+            Assert.Equal("Specified argument was out of the range of valid values. (Parameter 'Found 1 validation errors:\nError: FHIR field DeathLocationJurisdiction has value 'QQ', which is invalid for IJE field DSTATE.')", e.Message);
         }
 
         private string FixturePath(string filePath)